+++ steam_api/Steamworks/Interfaces/SteamNetworking/SteamNetworking005.cs
@@ -46,7 +46,7 @@ namespace SKYNET.Steamworks.Interfaces
             return SteamEmulator.SteamNetworking.AllowP2PPacketRelay(bAllow);
         }
 
-        public uint CreateListenSocket(IntPtr _, int nVirtualP2PPort, uint nIP, uint nPort, bool bAllowUseOfPacketRelay)
+        public uint CreateListenSocket(IntPtr _, int nVirtualP2PPort, uint nIP, ushort nPort, bool bAllowUseOfPacketRelay)
         {
             return SteamEmulator.SteamNetworking.CreateListenSocket(nVirtualP2PPort, nIP, nPort, bAllowUseOfPacketRelay);
         }
@@ -56,7 +56,7 @@ namespace SKYNET.Steamworks.Interfaces
             return SteamEmulator.SteamNetworking.CreateP2PConnectionSocket(steamIDTarget, nVirtualPort, nTimeoutSec, bAllowUseOfPacketRelay);
         }
 
-        public uint CreateConnectionSocket(IntPtr _, uint nIP, uint nPort, int nTimeoutSec)
+        public uint CreateConnectionSocket(IntPtr _, uint nIP, ushort nPort, int nTimeoutSec)
         {
             return SteamEmulator.SteamNetworking.CreateConnectionSocket(nIP, nPort, nTimeoutSec);
         }
@@ -76,32 +76,32 @@ namespace SKYNET.Steamworks.Interfaces
             return SteamEmulator.SteamNetworking.SendDataOnSocket(hSocket, pubData, cubData, bReliable);
         }
 
-        public bool IsDataAvailableOnSocket(IntPtr _, uint hSocket, uint pcubMsgSize)
+        public bool IsDataAvailableOnSocket(IntPtr _, uint hSocket, IntPtr pcubMsgSize)
         {
             return SteamEmulator.SteamNetworking.IsDataAvailableOnSocket(hSocket, pcubMsgSize);
         }
 
-        public bool RetrieveDataFromSocket(IntPtr _, uint hSocket, IntPtr pubDest, uint cubDest, uint pcubMsgSize)
+        public bool RetrieveDataFromSocket(IntPtr _, uint hSocket, IntPtr pubDest, uint cubDest, IntPtr pcubMsgSize)
         {
             return SteamEmulator.SteamNetworking.RetrieveDataFromSocket(hSocket, pubDest, cubDest, pcubMsgSize);
         }
 
-        public bool IsDataAvailable(IntPtr _, uint hListenSocket, uint pcubMsgSize, uint phSocket)
+        public bool IsDataAvailable(IntPtr _, uint hListenSocket, IntPtr pcubMsgSize, IntPtr phSocket)
         {
             return SteamEmulator.SteamNetworking.IsDataAvailable(hListenSocket, pcubMsgSize, phSocket);
         }
 
-        public bool RetrieveData(IntPtr _, uint hListenSocket, IntPtr pubDest, uint cubDest, uint pcubMsgSize, uint phSocket)
+        public bool RetrieveData(IntPtr _, uint hListenSocket, IntPtr pubDest, uint cubDest, IntPtr pcubMsgSize, IntPtr phSocket)
         {
             return SteamEmulator.SteamNetworking.RetrieveData(hListenSocket, pubDest, cubDest, pcubMsgSize, phSocket);
         }
 
-        public bool GetSocketInfo(IntPtr _, uint hSocket, ulong pSteamIDRemote, int peSocketStatus, uint punIPRemote, uint punPortRemote)
+        public bool GetSocketInfo(IntPtr _, uint hSocket, IntPtr pSteamIDRemote, IntPtr peSocketStatus, IntPtr punIPRemote, IntPtr punPortRemote)
         {
             return SteamEmulator.SteamNetworking.GetSocketInfo(hSocket, pSteamIDRemote, peSocketStatus, punIPRemote, punPortRemote);
         }
 
-        public bool GetListenSocketInfo(IntPtr _, uint hListenSocket, uint pnIP, uint pnPort)
+        public bool GetListenSocketInfo(IntPtr _, uint hListenSocket, IntPtr pnIP, IntPtr pnPort)
         {
             return SteamEmulator.SteamNetworking.GetListenSocketInfo(hListenSocket, pnIP, pnPort);
         }
